server/requirements.txt
requests
flask
flask-cors
psycopg2-binary
sqlalchemy
flask-marshmallow
marshmallow